+++ src/pages/post-detail/api/useGetPostDetail.ts
@@ -1,7 +1,6 @@
-import { useAxios as Axios } from "../../../shared/utils/useAxios";
+import axiosInstance from "../../../shared/utils/useAxios";
 
 export const useGetPostDetail = async (postId: number, onError: () => void) => {
-  const { axiosInstance } = Axios();
   try {
     const response = await axiosInstance.get(`/hf/posts/${postId}`);
     return response.data.content;
