+++ src/pages/post-register/api/usePostUpdate.ts
@@ -1,4 +1,4 @@
-import { useAxios as Axios } from "../../../shared/utils/useAxios";
+import axiosInstance from "../../../shared/utils/useAxios";
 
 interface UpdateProps {
   postId: number;
@@ -9,7 +9,6 @@ interface UpdateProps {
 }
 
 export const usePostUpdate = async (postData: UpdateProps) => {
-  const { axiosInstance } = Axios();
   try {
     const response = await axiosInstance.patch(
       `/hf/posts/${postData.postId}`,
