+++ src/shared/api/useGetSearchData.ts
@@ -1,7 +1,6 @@
-import { useAxios as Axios } from "../utils/useAxios";
+import axiosInstance from "../utils/useAxios";
 
 export const useGetSearchData = async (currentPage: number) => {
-  const { axiosInstance } = Axios();
   const urlParams = new URLSearchParams(window.location.search);
   const queryParams = Object.fromEntries(urlParams.entries());
