+++ src/shared/api/useGetUserData.ts
@@ -1,7 +1,7 @@
-import { useAxios as Axios } from "../utils/useAxios";
+import axiosInstance from "../utils/useAxios";
 
 export const useGetUserData = async (memberId: number) => {
-  const { axiosInstance } = Axios();
+  // const { axiosInstance } = Axios();
   try {
     const response = await axiosInstance(`/hf/members/${memberId}`);
     return response.data.content;
