+++ src/widgets/post-detail/comment-list/api/useGetComments.ts
@@ -1,8 +1,7 @@
-import { useAxios as Axios } from "../../../../shared/utils/useAxios";
+import axiosInstance from "../../../../shared/utils/useAxios";
 import Cookies from "js-cookie";
 
 export const useGetComments = async (postId: number) => {
-  const { axiosInstance } = Axios();
   const accessToken = Cookies.get("access_token");
   try {
     const response = await axiosInstance.get(`/hf/posts/${postId}/comments`, {
