+++ src/widgets/post-detail/post-content/api/useDeletePost.ts
@@ -1,7 +1,6 @@
-import { useAxios as Axios } from "../../../../shared/utils/useAxios";
+import axiosInstance from "../../../../shared/utils/useAxios";
 
 export const useDeletePost = async (postId: number) => {
-  const { axiosInstance } = Axios();
   try {
     const response = await axiosInstance.delete(`/hf/posts/${postId}`, {
     });
