+++ src/widgets/post-detail/post-content/api/useDeletePostLike.ts
@@ -1,8 +1,7 @@
-import { useAxios as Axios } from "../../../../shared/utils/useAxios";
+import axiosInstance from "../../../../shared/utils/useAxios";
 import Cookies from "js-cookie";
 
 export const useDeletePostLike = async (likeId: number) => {
-  const { axiosInstance } = Axios();
   const accessToken = Cookies.get("access_token");
   try {
     const response = await axiosInstance.delete(`/hf/likes/${likeId}`, {
