+++ src/widgets/post-detail/post-content/api/useGetLike.ts
@@ -1,11 +1,10 @@
-import { useAxios as Axios } from "../../../../shared/utils/useAxios";
+import axiosInstance from "../../../../shared/utils/useAxios";
 import Cookies from "js-cookie";
 
 export const useGetLike = async (
   postId: number,
   memberId: number | undefined
 ) => {
-  const { axiosInstance } = Axios();
   const accessToken = Cookies.get("access_token");
   try {
     const response = await axiosInstance.get(`/hf/posts/${postId}/likes`, {
